requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
PyMySQL==1.1.0
cryptography==41.0.3

# OCR & Image Processing
pytesseract==0.3.10
opencv-python==4.8.0.76
Pillow==10.0.0
pdf2image==1.16.3

# Data Processing
numpy==1.24.3
python-dateutil==2.8.2

# Forms & Validation
email-validator==2.0.0
WTForms==3.0.1
Flask-WTF==1.1.1

# Utilities
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3

# Testing (optional)
pytest==7.4.0
pytest-cov==4.1.0

# Production Server (optional)
gunicorn==21.2.0
